% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addColumnDF.R
\name{addColumnDF}
\alias{addColumnDF}
\title{Dataframe Numeric Column Adder}
\usage{
addColumnDF(peaks.data)
}
\arguments{
\item{peaks.data}{the dataframe which you wish to add a column too}
}
\value{
the same dataframe with an added numeric column
}
\description{
function which adds a numeric column to a data frame
}
\examples{
newdataframe.data = addcolumnDF(peak.data)

}
