% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addColumnDFCharacter.R
\name{addColumnDF_character}
\alias{addColumnDF_character}
\title{Dataframe Charcter Column Adder}
\usage{
addColumnDF_character(peaks.data)
}
\arguments{
\item{peaks.data}{the dataframe which you wish to add a column too}
}
\value{
the same dataframe with an added character column
}
\description{
function which adds a character column to a data frame
}
\examples{
newdataframe.data = addcolumnDFCharacter(peak.data)

}
