% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDF.R
\name{buildDF}
\alias{buildDF}
\title{Peak Dataframe Builder}
\usage{
buildDF(uniquePeaks, peaks.data)
}
\arguments{
\item{uniquePeaks}{the vector containing uniquepeaks from a Conn file}

\item{peaks.data}{the dataframe which you wish to add a column too}
}
\value{
the same dataframe now filled with peak data
}
\description{
function which fills empty data frame with given peak data
}
\examples{
peak.data = buildDF(uniquePeaks, peak.data)

}
