% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDFMatrix.R
\name{buildDF_matrix}
\alias{buildDF_matrix}
\title{Peak Bed Dataframe Builder}
\usage{
buildDF_matrix(uniquePeaks)
}
\arguments{
\item{uniquePeaks}{a vector containing all uniquepeaks}
}
\value{
a new dataframe filled with given peak data
}
\description{
fills an empty data frame with only peak data
}
\examples{
peak.data = buildDF_bed(uniquePeaks)

}
