% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciceroSplit.R
\name{cicero_split}
\alias{cicero_split}
\title{Cicero Output Splitter}
\usage{
cicero_split(cicString)
}
\arguments{
\item{cicString}{a string in the form Chr'_'''''''_''''''''}
}
\value{
a vector containg start, end, and chromosone
}
\description{
takes cicero output string and returns a vector with: end, start, chromosone
}
\examples{
vector = cicero_split(cicString)

}
