% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciceroSplitSpace.R
\name{cicero_split_space}
\alias{cicero_split_space}
\title{Cicero Output Splitter Space}
\usage{
cicero_split_space(cicString)
}
\arguments{
\item{cicString}{a string in the form Chr' ''''''' ''''''''}
}
\value{
a vector containg start, end, and chromosone
}
\description{
takes tab dilemited output string and returns a vector with: end, start, chromosone
}
\examples{
vector = cicero_split(cicString)

}
