% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareConns.R
\name{compareConns}
\alias{compareConns}
\title{Connection Comparision}
\usage{
compareConns(WT_conns, KO_conns)
}
\arguments{
\item{WT_conns}{a conns file}

\item{KO_conns}{a conns file}
}
\value{
a dataframe containing the results of the comparision
}
\description{
compares two Conns files by peak similarity and total connections
}
\examples{
results.data = compareConns(WT_conns, KO_conns)

}
