% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connBlender.R
\name{connBlender}
\alias{connBlender}
\title{Blend two conns files}
\usage{
connBlender(KO_conns, WT_conns)
}
\arguments{
\item{KO_conns}{the first dataframe you want to combine}

\item{WT_conns}{the second dataframe you want to combine}
}
\value{
a blended dataframe
}
\description{
Takes in two conns files and combines like peaks and adds peaks not included in each others
}
\examples{
loop.data = sortConnections(loop.data, bed)

}
