% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connBlendInt.R
\name{connblendInt}
\alias{connblendInt}
\title{Calculate Total Connections}
\usage{
connblendInt(SUMpeaksblend.data, KO_conns, thresh)
}
\arguments{
\item{SUMpeaksblend.data}{a data frame containting all the unique peaks}

\item{KO_conns}{a conns file}

\item{thresh}{the decimal certainty above which you will accept peaks}
}
\value{
a dataframe containing the results of the conn totaling
}
\description{
Calculate # of connections for each peak in a given conns file over a threshold
}
\examples{
SUMpeaksKO.data = connblendInt(SUMpeaksblend.data, KO_conns, thresh)

}
