% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connBlendWeight.R
\name{connblendweight}
\alias{connblendweight}
\title{Calculate Total Connections Weighted}
\usage{
connblendweight(SUMpeaksblend.data, KO_conns, thresh)
}
\arguments{
\item{SUMpeaksblend.data}{a data frame containting all the unique peaks}

\item{KO_conns}{a conns file}

\item{thresh}{the decimal certainty above which you will accept peaks}
}
\value{
a dataframe containing the results of the conn weights
}
\description{
Calculate total weight of connections for each peak in a given conns file over a threshold
}
\examples{
SUMpeaksKO.data = connblendInt(SUMpeaksblend.data, KO_conns, thresh)

}
