% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connCreator.R
\name{connCreator}
\alias{connCreator}
\title{Create a conns files}
\usage{
connCreator(WT_cicero_cds, number, mm10.chr)
}
\arguments{
\item{WT_cicero_cds}{the cds you want to convert to a conns file}

\item{number}{the chromosone you want a conns file for}

\item{mm10.chr}{the genome file you are referencing to create the conns file}
}
\value{
a conns file
}
\description{
Takes in a cds files and creates a conns files
}
\examples{
KO_conns = connCreator(T_cicero_cds, number, mm10.chr)

}
