% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connInteger.R
\name{connInteger}
\alias{connInteger}
\title{Create a conns files and then analyzes them}
\usage{
connInteger(WT_cicero_cds, KO_cicero_cds, vector, mm10.chr, namefirst,
  namesecond, thresh, changePerc, viablethresh)
}
\arguments{
\item{WT_cicero_cds}{the first cds you want to convert to a conns file}

\item{KO_cicero_cds}{the second cds you want to convert to a conns file}

\item{vector}{the vector containing which chromosones you are interested in}

\item{mm10.chr}{the genome file you are referencing to create the conns files}

\item{namefirst}{the label you want to give to the first conns file, i.e. "A"}

\item{namesecond}{the label you want to give to the second conns file, i.e. "B"}

\item{thresh}{the threshold for peaks you are interested in}

\item{changePerc}{the percent difference between Ko and WT conns you want to use as a threshold}

\item{viablethresh}{the threshold for peaks you consider viable}
}
\value{
a completed, sorted, and eleaborated dataframe
}
\description{
Takes in a cds files and turns them into conns files and then analyzes them with an integer weighting
}
\examples{
SUM.data = connWeight(WT_cicero_cds, KO_cicero_cds, vector, mm10.chr, namefirst, namesecond, thresh, changPerc)

}
