% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connWeightDirect.R
\name{connWeightDirect}
\alias{connWeightDirect}
\title{Analyzes two conns files}
\usage{
connWeightDirect(WT_conns, KO_conns, namefirst, namesecond, thresh,
  changePerc, viablethresh)
}
\arguments{
\item{WT_conns}{the first conns file}

\item{KO_conns}{the second conns file}

\item{namefirst}{the label you want to give to the first conns file, i.e. "A"}

\item{namesecond}{the label you want to give to the second conns file, i.e. "B"}

\item{thresh}{the threshold for peaks you are interested in}

\item{changePerc}{the percent difference between Ko and WT conns you want to use as a threshold}

\item{viablethresh}{the threshold for peaks you consider viable}

\item{mm10.chr}{the genome file you are referencing to create the conns files}
}
\value{
a completed, sorted, and eleaborated dataframe
}
\description{
Takes in two conns files and analyzes them fully with normal weighting
}
\examples{
SUM.data = connWeight(WT_conns, KO_conns, namefirst, namesecond, thresh, changPerc)

}
