% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDF.R
\name{createDF}
\alias{createDF}
\title{Create Empty Peak Dataframe}
\usage{
createDF(vectorsize)
}
\arguments{
\item{vectorsize}{the length you want the dataframe to be}
}
\value{
an empty dataframe of the correct length
}
\description{
Creates an empty peak dataframe of a given length
}
\examples{
peak.data = createDF(400)

}
