% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillMatrixRed.R
\name{fillMatrixRed}
\alias{fillMatrixRed}
\title{Fill A Matrix With Conns}
\usage{
fillMatrixRed(KO_conns, emptymatrix, peaksVector, thresh)
}
\arguments{
\item{KO_conns}{a conns file}

\item{emptymatrix}{an empty matrix of the same dimension as the peaks vector}

\item{peaksVector}{a vector of the peak bins that the conns will be sorted into}

\item{thresh}{the threshold over which a connection will be accepted}
}
\value{
a completed matrix filled with connsInt
}
\description{
Fills a graphable matrix based on a given peakvector
}
\examples{
matrix = fillMatrixRed(KO_conns, emptymatrix, peaksVector, thresh)

}
