% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCoord.R
\name{findCoord}
\alias{findCoord}
\title{Finds Place of Peak Along PeakVetor}
\usage{
findCoord(connsString, peaksVector, startSearch)
}
\arguments{
\item{peaksVector}{the numerically sorted list of peaks you're searching through}

\item{startSearch}{the number guess of where the peak may be near too}

\item{connString}{the conn format peak you are looking to locate}
}
\value{
a number corresponding to it's location
}
\description{
given a starting location and a list of peaks, finds where a given connString would fit (i.e. a coordinate)
}
\examples{
location = findCoord(connString, peaksVector, startSearch)

}
