% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listSplit.R
\name{list_split}
\alias{list_split}
\title{Create Peak Dataframe from List}
\usage{
list_split(cicString)
}
\arguments{
\item{cicString}{the string containing a list of peaks}
}
\value{
a peak dataframe
}
\description{
given a comma seperated list of peaks, creates a peak dataframe
}
\examples{
peaks.data = list_string(cicString)

}
