% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locaterBelow.R
\name{locaterBelow}
\alias{locaterBelow}
\title{Returns a Location Based on A Given Coordinate}
\usage{
locaterBelow(uniquepeaks, location)
}
\arguments{
\item{uniquepeaks}{a vector containing all uniweu peaks in a conns file}

\item{location}{the coordiante you are conserned with}
}
\value{
a integer coordinate
}
\description{
returns the position along a unique peak vector whcih coresponds to a given location along a chromosone
}
\examples{
locater = locaterBelow(uniquepeaks, location)

}
