% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakBlend.R
\name{peakBlend}
\alias{peakBlend}
\title{Peak Blender}
\usage{
peakBlend(SUMpeaks.data)
}
\arguments{
\item{SUMpeaks.data}{a peak dataframe}
}
\value{
a new peak dataframe with combined peaks at overlapping areas
}
\description{
combines overlapping peaks in a given peak dataframe
}
\examples{
peaks.data = peakBlend(SUMpeaks.data)

}
