% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaksCut.R
\name{peaksCut}
\alias{peaksCut}
\title{Peak Cutter and Reducer}
\usage{
peaksCut(uniquepeaks, dim, startNum, endNum)
}
\arguments{
\item{uniquepeaks}{a vector containing all the unique files in the conn file used to generate the simple matrix}

\item{dim}{the amount of peaks you want in each bin}

\item{startNum}{the start coordinate of your desired region}

\item{endNum}{the end coordinate of your desired region}
}
\value{
a dataframe containing the peak information of your reduced peaks
}
\description{
creates a dataframe which mirrors a reduced matrix to help in interpretation
}
\examples{
peaks.data = peaksCut(uniquepeaks, 70, 1200, 300000000)

}
