% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigPeakFinder.R
\name{sigPeakFinder}
\alias{sigPeakFinder}
\title{Signifigant Peak Finder}
\usage{
sigPeakFinder(SUM.data, thresh, percent)
}
\arguments{
\item{SUM.data}{a peak file containing two conn files connection data}

\item{thresh}{the amount of total connections in at least one conn file which the user determines is signifigant, only peaks above this will be taken}

\item{percent}{the percent change in amount of connection between each conn file which the user deems is signifigant}
}
\value{
a dataframe containing the peak information of your signifigant peaks
}
\description{
creates a dataframe which mirrors a reduced matrix to help in interpretation
}
\examples{
peaks.data = sigPeakFinder(SUM.data, 20, 50)

}
