% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleMatrixEven.R
\name{simpleMatrixEven}
\alias{simpleMatrixEven}
\title{Creates a Reduced Graphable Matrix of Conns values}
\usage{
simpleMatrixEven(KO_conns, dim, thresh, startNum, endNum)
}
\arguments{
\item{KO_conns}{a conns file}

\item{dim}{the amount of peaks you want in each bin}

\item{startNum}{the start coordinate of your desired region}

\item{endNum}{the end coordinate of your desired region}
}
\value{
a reduced matrix filled with connsInt
}
\description{
Creates a reduced graphable matrix based on a given conns file
}
\examples{
matrix = simpleMatrix(KO_conns, 70, 1200, 300000000)

}
