% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortConnections.R
\name{sortConnections}
\alias{sortConnections}
\title{Sort Connections of a completed dataframe}
\usage{
sortConnections(loop.data, bed)
}
\arguments{
\item{loop.data}{the generated dataframe whose connections you want to sort}

\item{bed}{the peaks you want to search for and keep within the conns}
}
\value{
the loop.data dataframe with newly sorted peaks
}
\description{
Takes a completed dataframe with KO and WT conns and sorts them based on a given bed file containing peaks
}
\examples{
loop.data = sortConnections(loop.data, bed)

}
