% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumPeaksName.R
\name{sumPeaksName}
\alias{sumPeaksName}
\title{Peak DataFrame Combination}
\usage{
sumPeaksName(WTpeaks.data, KOpeaks.data, firstString, secondString)
}
\arguments{
\item{WTpeaks.data}{a peak file containing the peaks of one conn file}

\item{KOpeaks.data}{a peak file containing the peaks of one conn file}

\item{firstString}{the character you want attached with the first conns file}

\item{secondString}{the character you want attached with the second conns file}
}
\value{
a dataframe containing the peak information of your signifigant peaks
}
\description{
creates a dataframe which adds two peak dataframe by combining like and independant peaks and marks each as such
}
\examples{
peaks.data = sigPeakFinder(SUM.data, 20, 50)

}
