% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumToPeaks.R
\name{sumtoPeaks}
\alias{sumtoPeaks}
\title{Dataframe to Peak Vector}
\usage{
sumtoPeaks(SUM.data)
}
\arguments{
\item{SUM.data}{a peak dataframe}
}
\value{
a vector containing the peaks
}
\description{
turns a sum peak dataframe back into a peak vector
}
\examples{
vector = sumtoPeaks(SUM.data)

}
