% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniquePeak.R
\name{unique_peak}
\alias{unique_peak}
\title{Unique Peak Vector}
\usage{
unique_peak(conns)
}
\arguments{
\item{conns}{a conns file}
}
\value{
a vector containing the peaks
}
\description{
takes every unique peak from a conns file and returns a vector of them
}
\examples{
vector = unique_peak(conns)

}
