% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connInteger.R
\name{connInteger}
\alias{connInteger}
\title{Create a conns files}
\usage{
connInteger(WT_cicero_cds, KO_cicero_cds, vector, mm10.chr, namefirst,
  namesecond)
}
\arguments{
\item{WT_cicero_cds}{the first cds you want to convert to a conns file}

\item{KO_cicero_cds}{the second cds you want to convert to a conns file}

\item{vector}{the vector containing which chromosones you are interested in}

\item{mm10.chr}{the genome file you are referencing to create the conns files}

\item{namefirst}{the label you want to give to the first conns file, i.e. "A"}

\item{namesecond}{the label you want to give to the second conns file, i.e. "B"}
}
\value{
a blended dataframe
}
\description{
Takes in a cds files and creates a conns files
}
\examples{
loop.data = sortConnections(loop.data, bed)

}
